requirements.txt
# Alert Triage System - Production Dependencies

# Core framework dependencies
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0
pydantic>=2.5.0
pyyaml>=6.0.1
structlog>=23.2.0

# Data processing and analysis
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Database and caching
asyncpg>=0.29.0
aioredis>=2.0.1
sqlalchemy>=2.0.0

# HTTP client and networking
httpx>=0.25.0
dnspython>=2.4.0

# Security and cryptography
cryptography>=41.0.0
bcrypt>=4.0.0
pyjwt>=2.8.0

# Monitoring and metrics
prometheus-client>=0.19.0
psutil>=5.9.0

# Date and time handling
python-dateutil>=2.8.0
pytz>=2023.3

# Serialization and data formats
msgpack>=1.0.7
orjson>=3.9.0

# Logging and debugging
colorlog>=6.8.0
rich>=13.7.0

# Configuration management
python-dotenv>=1.0.0
click>=8.1.0

# Development and testing (optional in production)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.7.0

# Optional ML dependencies
# torch>=2.1.0
# transformers>=4.35.0
# xgboost>=2.0.0

# Optional threat intelligence
# yara-python>=4.3.0
# python-magic>=0.4.27

# Optional integrations
# slack-sdk>=3.23.0
# elasticsearch>=8.11.0
# kafka-python>=2.0.2